requirements.txt
# AI Agent Dependencies for SentinelAI
crewai>=1.11.0
langchain-groq>=0.2.9
langchain>=0.3.8
langchain-community>=0.3.9
python-dotenv>=1.0.1
requests>=2.32.3
dnspython>=2.7.0
aiohttp>=3.11.7
pydantic>=2.10.3
litellm>=1.57.0